Fixes #5299. Removes some code that is never executed.
Note: A number of the changes in the patch were in code that is synced with external code, so I did not make those changes since the current code is harmless and we want to make sure that future syncs remain simple. Other code was obviously left in for future extension, and this was also left in and commented.

git-svn-id: http://svn.mythtv.org/trunk@17268 4b91aca1-2a1f-0410-a73a-a78f2df169e8