Fix the visualisers not working in fullscreen mode in MythMusic since the switch to Qt4.
In Qt4 calling setUpdatesEnabled(false) disables updates for both the widget
and all its children which appears to differ from the Qt3 version which just
disabled updates for the calling widget.

Also fixes things so the visualiser widget doesn't use a QPainter outside
of a paint event which was causing lots of warning messages.

Fixes #5361.



git-svn-id: http://svn.mythtv.org/trunk@17344 4b91aca1-2a1f-0410-a73a-a78f2df169e8