Fix a segfault in MythControls when setting a keybinding.
Not sure if this is a Qt4 thing or due to a compiler change, the same code
worked in the fixes branch. The compiler complains about taking the address of
a temporary in ActionSet::GetActions() and sure enough the first time the
reference is used it segfaults.

Just changed it to return by value, a little less efficient but the list is
usually empty or contains only one value anyway.

git-svn-id: http://svn.mythtv.org/trunk@17854 4b91aca1-2a1f-0410-a73a-a78f2df169e8