Allow <area> tags to hold percentage values.
This feature has only been added to ease the creation of scalable popup dialogs and lists. It is not envisioned as a replacement for all absolute values in a theme. That approach would be not recommended, not least because of the loss of pixel precision in the conversion between percentages and integer coordinates.



git-svn-id: http://svn.mythtv.org/trunk@18014 4b91aca1-2a1f-0410-a73a-a78f2df169e8