Fixes a number of segfaults on error conditions and exits the TV event loop when TV::errored is set.
The data structure for QStringList has changed in Qt4 and will now segfault if you try to read past the end of the string list, so we need to check the length of the list returned in remoteencoder.

This usually only happens when the backend dies, so this isn't a regular occurance; but it can also happen on simple network timeouts.



git-svn-id: http://svn.mythtv.org/trunk@18098 4b91aca1-2a1f-0410-a73a-a78f2df169e8