Add a new mythui widget for displaying HTML documents based on QtWebKit.
QtWebKit is only available in Qt4.4.0 or later but because we wish to continue
to support Qt4.3.x until most of the major distros are shipping it there is
two versions of the widget, one for Qt4.3.x which just displays a "This 
feature requires Qt4.4.0 or later" message and a >=Qt4.4.0 version which
provides the full featured widget.

This means you can still use Myth with Qt4.3.x but any new features, or old
ones converted to use the new widget, will have reduced functionality until
you upgrade to Qt4.4.x.

For developers wishing to use the new widget there's some doxygen comments in
mythuiwebbrowser.cpp.


git-svn-id: http://svn.mythtv.org/trunk@18140 4b91aca1-2a1f-0410-a73a-a78f2df169e8