Merge MythUI port of MythVideo back to trunk.
I expect there to be some problems, this was more than a simple conversion, a lot of changes were made to the UI and refactoring was done, 6000 lines of code and theme xml were eliminated. I would not update if MythVideo is mission-critical to you, give it a couple of weeks for the bugs to be shaken out first.

This change may break some themes.

Refs #12



git-svn-id: http://svn.mythtv.org/trunk@18389 4b91aca1-2a1f-0410-a73a-a78f2df169e8