Change the format of XML used to define a textedit. This is arguably a less concise form and not as pretty as what it replaces, however it's a small price to pay for increased power it gives to themers and the extensibility it provides. This format has already been adopted for MythUIButtonList and MythUIProgressBar.
In brief it allows any formatting and styling to be applied to individual components of the widget. It can have animated backgrounds and cursors, colour cycling text and unusual shapes etc. New features added to component widgets (MythUIText, MythUIImage, MythUIStateType) become instantly available without the need to modify or extend this widget. The extra xml is balanced by inheritance.

Took the opportunity to add different backgrounds depending on state, which had been missing until now. It should be easier to tell which TextEdit is selected.



git-svn-id: http://svn.mythtv.org/trunk@18522 4b91aca1-2a1f-0410-a73a-a78f2df169e8