Refs #5596. Gets rid of Q3SimpleRichText.
Note the rendering is by no means as pretty with the new Qt4 rich text handler. It can be, but now we need to use CSS for the same effect, and I did not bother to make this pixel perfect since this should soon be replaced with a MythUI component.

git-svn-id: http://svn.mythtv.org/trunk@18650 4b91aca1-2a1f-0410-a73a-a78f2df169e8