Refs #5606. Turns off qt3 compat for libmythtv by default.
There are two things that are now disabled by default because they have not yet been ported, dbox2 support and mheg support. These can be enabled by using --enable-dbox2 and --enable-mheg respectively when running ./configure.

Enabling either one will pull in the qt3 compat library. I'll be attaching a patch to #5606 shortly which does a partial port to Qt4 for MHEG, but someone with access to MHEG streams will need to complete the port.

I highly suggest doing a make distclean, but it should not be required in most cases.



git-svn-id: http://svn.mythtv.org/trunk@18762 4b91aca1-2a1f-0410-a73a-a78f2df169e8