Closes #5268, #5319
  * Allows for checkTimeZone() to not establish a connection to the MBE.
  * Changes the way checkTimeZone() works for SBEs
    * Create a dummy Monitor connection to the MBE which doesn't accept
      events and is closed after checkTimeZone() is done.
    * Only perform checkTimeZone() if the MBE is running, previously it
      enforced a startup order we've never done before.
  * Some minor cleanup I've had sitting around.

What I didn't do, make the tz checking happen on each SBE reconnect, doing
so would now be easy but seems a little overboard.

Note: This doesn't fix the issue originally reported in #5268. I haven't
found the other cause but it is likely the same issue (some function in
MythContext being called that establishes an event connection to the MBE).


git-svn-id: http://svn.mythtv.org/trunk@18792 4b91aca1-2a1f-0410-a73a-a78f2df169e8