Update the program details editor in MythArchive to MythUI.
Editing the description is a bit of pain in a single line edit but it's
all we've got.

There seems to a bug in MythUITextEdit that causes it to draw text out of the
text edit area if the text is too large for it.

It's currently impossible to use a remote to edit text.



git-svn-id: http://svn.mythtv.org/trunk@19181 4b91aca1-2a1f-0410-a73a-a78f2df169e8