Update VDPAU playback for the latest (180.16) driver/API.
I used the approach taken in the updated mplayer patches to calculate the number of references frames. This works but usually heavily overestimates the number required. I also tested the approach, hinted at in the mplayer patches, of using the h264 parser to give the actual requirement. This works well for most streams but fails with certain files (e.g. QuickTime/mov) and at least one file reported a varying number of reference frames (not sure if that is legal but haven't checked).


git-svn-id: http://svn.mythtv.org/trunk@19356 4b91aca1-2a1f-0410-a73a-a78f2df169e8