Refs #2287. Merges mythtv-vid to trunk.
There are a number of changes, probably the most user visible ones are PiP related, but there is also groundwork laid for OpenGL video rendering on OSX and MS Windows, for aspect ratio & zoom on OSX, and a number of operations are more responsive.

Note: Even though this changeset has received a fair bit of testing, it is a 1.1MB patch and there are bound to be things we overlooked. Please report any compilation problems to the mythtv-dev mailing list and open tickets with a repro description, theme used and backtrace for any segfaults or deadlocks.

git-svn-id: http://svn.mythtv.org/trunk@19417 4b91aca1-2a1f-0410-a73a-a78f2df169e8