VDPAU PIP. Re-work and re-enable vdpau pip following the -vid merge.
Allows up to 4 pips (consistent with XVideo and OpenGL). As current VDPAU implementations seem to be limited to 4 layers for compositing, and we need 1 for the OSD, just use 1 layer for all pips and blend each pip into that layer.
N.B. Just to be clear, pips are NOT decoded in hardware.


git-svn-id: http://svn.mythtv.org/trunk@19622 4b91aca1-2a1f-0410-a73a-a78f2df169e8