Port the Program Details screen to MythUI.
Notes:-

This uses the MythUIWebBrowser widget that only works on Qt4.4.0 or later.

The info is now split over two pages. In the default theme the first page
shows the information that most users want to see. The second page shows
everything that is on the first plus some more advanced stuff. 'INFO'
switches between screens.

The information that is shown on each page is fully themeable by just editing
progdetails_page{1,2}.html. The html files are searched for in the same manor
as all UI files so themers can add there own versions if required. They should
at least change the font colors.

I've removed the old Qt screen completely, it's been broken since the switch
to Qt4 anyway. To keep everyone happy until everything is switched to MythUI
I've used a temporary hack to enable the new screen to be shown on top of a Qt
one. The hack does cause a slight flicker when switching between Qt and MythUI
screens but at least it keeps things working until everything is switched.

This screen should probably be moved to mythfrontend once all the screens that
call it are moved out of libmythtv.



git-svn-id: http://svn.mythtv.org/trunk@19643 4b91aca1-2a1f-0410-a73a-a78f2df169e8