Convert the Program Finder screen to MythUI.
You can now enter some search text to filter the program list to show only
programs that contain the text you enter.

There is a new pop menu for easy access to everything that was previously
only available though key strokes.

The area reserved for the video has gone, it was never used for its intended
purpose and just wasted screen space. It can always be added back later if we
ever get a working video widget.

The Japanese and Hebrew versions are untested.

The default-wide theme still needs to be updated.


git-svn-id: http://svn.mythtv.org/trunk@20138 4b91aca1-2a1f-0410-a73a-a78f2df169e8